refactor(nvim): move vim-clipper submodule from next to main
On the "next" branch I was justifying using OSC 52 on the basis of the
trick used in:

https://github.com/wincent/wincent/blob/c85890b0dada7ad1/aspects/dotfiles/files/.zsh/bin/clip

As commit c85890b0dada7ad1 says:

    refactor(dotfiles): switch back to using OSC 52 for `clip` when in Kitty

    Works fine on Kitty's current `master`, as I had hoped it would.

Well, I ended up backing that out in f0535c6c84c90104 ("feat: get dual
local/remote clipboards working on Linux box", Jul 25 2022). So, go back
to the old ways. If nothing else, because I was getting errors on trying
to write to `/dev/tty`.

* aspects/nvim/files/.config/nvim/pack/bundle/opt/vim-clipper b6e5b69...e239712 (1):
  < feat: use OSC 52 sequence when in kitty