chore(nvim): set 'pumheight' to 20
Default is zero, which means the popup menu will fill as many lines as
needed/possible, but I find that if it is more than about 20 then I am
not going to come even close to scanning that many entries and it just
makes it more likely that my eyes will have to jump farther to see the
first item in the menu (ie. if I am near the bottom of the viewport and
the first item in the list is way near the top).