fix(dotfiles): make C standard libary man pages findable on macOS again
(Not just C standard library, but a bunch of others too.)

This seems to be related to kitty, where I noticed that stuff like `man
_exit` (etc) stopped working.

In iTerm, `echo $MANPATH` prints nothing, and `man` works.

In Kitty, `echo $MANPATH` was printing this (formatted for readability):

    /usr/share/man
    /usr/local/share/man
    /opt/X11/share/man
    /Applications/kitty.app/Contents/Resources/man

Perhaps there's some kind of default search path that is used when
`$MANPATH` is not set, but when it's set, the defaults are not
consulted? FWIW, I see this in `/etc/manpaths`:

    /usr/share/man
    /usr/local/share/man

and this inside `/etc/man.conf`:

    MANPATH	/usr/share/man
    MANPATH	/usr/local/share/man
    MANPATH	/usr/X11/man
    MANPATH	/Library/Apple/usr/share/man

`/etc/manpaths.d` contains only `40-XQuartz`, which contains this:

    /opt/X11/share/man

My understanding is that when `MANPATH` is an empty string, the four
items listed in `/etc/man.conf` end up being what gets searched.

Of all of the directories listed above:

    [exists, 2 pages]      /Applications/kitty.app/Contents/Resources/man
    [exists, 9 pages]      /Library/Apple/usr/share/man
    [exists, 4019 pages]   /opt/X11/share/man
    [missing]              /usr/X11/man
    [exists, 13682 pages]  /usr/local/share/man
    [exists, 3318 pages]   /usr/share/man

Anyway, after the change in this commit, my MANPATH now looks like
this (in kitty) - again, formatted for readability, and with counts in
brackets:

    [exists, 3318 pages]   /usr/share/man
    [exists, 13682 pages]  /usr/local/share/man
    [exists, 9 pages]      /Library/Apple/usr/share/man
    [exists, 4019 pages]   /opt/X11/share/man
    [exists, 2 pages]      /Applications/kitty.app/Contents/Resources/man
    [exists, 16028 pages]  /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX.sdk/usr/share/man
    [exists, 231 pages]    /Applications/Xcode.app/Contents/Developer/usr/share/man
    [exists, 86 files]     /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/share/man

and like this in iTerm:

    [exists, 3318 pages]   /usr/share/man
    [exists, 13682 pages]  /usr/local/share/man
    [exists, 9 pages]      /Library/Apple/usr/share/man
    [exists, 16028 pages]  /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX.sdk/usr/share/man
    [exists, 231 pages]    /Applications/Xcode.app/Contents/Developer/usr/share/man
    [exists, 86 files]     /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/share/man

ie. both lists are the same, except kitty adds "/opt/X11/share/man" and
(of course) "/Applications/kitty.app/Contents/Resources/man".

Overall, I'm not thrilled with the hard-codedness of this all,
especially given that mechanisms like `/etc/man.conf` and
`/etc/manpaths`/`/etc/manpaths.d/` are supposed to provide a good
default for us, but at least the manpages are unbroken now.

How many duplicates do we have in all of this? My first attempt at this
was flawed because it ended up printing out (unique) absolute path
names:

    find ${^${(s/:/)MANPATH}} -type f | wc -l # 22621 files
    find ${^${(s/:/)MANPATH}} -type f | sort | uniq | wc -l # 22621 files

Revised attempt (much, much slower, because it is forking `basename`
tens of thousands of times along the way):

    find ${^${(s/:/)MANPATH}} -type f -exec basename {} \; | wc -l # 22621 files again
    find ${^${(s/:/)MANPATH}} -type f -exec basename {} \; | sort | uniq | wc -l # 19636 files

So, almost 3 thousand duplicates, which we can print like this:

    typeset -A COUNTS

    for F in $(find ${^${(s/:/)MANPATH}} -type f -exec basename {} \;); do
      if (( COUNTS[$F] )); then
        COUNTS[$F]+=1
      else
        COUNTS[$F]=1
      fi
    done

    for F in "${(@k)COUNTS}"; do
      if (( COUNTS[$F] > 1 )); then
        echo "$F"
      fi
    done

Eyeballing a sampling of those, they look to be a wide variety of
things, too varied to summarize here. I could dig deeper to see where
the dupes are coming from, compare the contents etc, but I am simply too
lazy.

Got the `xcode-select --show-manpaths` hint from:

- https://stackoverflow.com/a/76744505/2103996