chore(dotfiles): turn on `autoStash`
I've delayed doing this for years because of the warning in the man
page:

> However, use with care: the final stash application after a successful
> merge might result in non-trivial conflicts.

But I got sick of having this experience:

    $ git pull # fails...
    error: Your local changes to the following files would be
    overwritten by merge:
            ...
    Please commit your changes or stash them before you merge.
    $ git ff # works without issue

So, feck it. Turning it on for rebase and merge, and removing
`--autostash` from my `ff` alias because it's no longer needed.

The output during `pull` etc is hopefully enough to help me unfuck
myself in the event of a conflict:

    From ...
     * branch              main       -> FETCH_HEAD
    Updating aced4999..349589f6
    Created autostash: ff652f77
    Fast-forward
    ...
     6 files changed, 101 insertions(+), 71 deletions(-)
    Applied autostash.

Along the way, I turned a number of config options to camelCase because
that's what the `git-config` man page started doing some time ago (I
don't know exactly when and am too lazy to look it up).