perf(nvim): use experimental loader
This was added in:

- https://github.com/neovim/neovim/pull/22668 (March 26, 2023)

And will likely be on by default in the future:

- https://github.com/neovim/neovim/pull/24044

Highly scientific testing done with `--startuptime` option:

1. First run, before this commit: 197.184ms (no bytecode cache)
2. Second run, after this commit: 196.709ms (writing to bytecode cache)
3. Third run: 145.194ms (reading from bytecode cache)
4. Fourth run, to double-check result: 156.129ms
5. Fifth run, to triple-check result: 144.645ms

Free perf, so may as well grab it. There may be a bit of a learning
curve in here around when/how to invalidate the cache, or how to
troubleshoot stale cache issues.

For now, `:echo stdpath('cache')` reveals that the cache is stored in
`~/.cache/nvim`, and that PR linked above reveals that the cached
bytecode lives in the `luac` subdirectory (mine currently has 247 files
in it).