feat(nvim): set up shellbot
I'm not automating the build yet (it is a single command to run), as I
am not sure whether this will be a permanent part of my set-up or not.

It will be pretty easy to do so, though, later on.

Note I had to set up a custom mapping to call `ChatGPTSubmit`
(unfortunately, a global), because `<C-CR>` doesn't work inside tmux.
Works fine outside of it though. Anyway, using `<M-CR>` instead. I was
going to just use a `<CR>` normal mode mapping, but the plugin blows it
away so I can't do that.