fix(nvim): suppress "Do you need to configure your work environment" spam
As suggested here:

- https://smarttech101.com/nvim-lsp-configure-language-servers-shortcuts-highlights/

> If you don’t know what LÖVE/ Luv, and Luassert are, chances are you
> don’t need them. So, disable them permanently

Apparently they are a game engine and an assertion library,
respectively.

Note that lua-language-server only a few days ago changed
`checkThirdParty` from a boolean to a string enum in:

- https://github.com/LuaLS/lua-language-server/pull/2354

but backwards compatibility is maintained (`false` should be equivalent
to `"Disable"`), so I'm going with `false` to have this work in as many
places as possible.