fix(nvim): improve handling of spell-related settings
This is probably not a minimal fix, as I tried a few things along the
way before arriving at this point, while trying to fix:

- https://github.com/wincent/wincent/issues/143

but I think the end result is probably worth keeping in its full form.

The motivating problem was that I wasn't seeing `'spell'` being set on
Markdown files. As noted on that issue, commenting out the
problem-plagued `ownsyntax` stuff made the problem go away.

I observed that on opening vim with `nvim CHANGELOG.md`, for example,
we'd get focus events that would end up clearing the `'spell'` setting
because we hadn't captured the value before calling `:ownsyntax on`. So
I initially added an additional `capture()` call before the first
invocation of `:ownsyntax on`.

But then I saw that open opening a new file in the same buffer, say,
`fig/main.ts`, that it still had `'spell'` set. So I added per-filetype
capturing, along with a `FileType` autocmd in order to make sure we got
the initial values for any given filetype.

That wasn't enough however, because the settings was evidently getting
applied to the window instead of the buffer, so I went down the rabbit
whole of `:setlocal` versus `vim.opt_local`, and the corresponding
low-level APIs. In the end `vim.api.nvim_set_option_value` does what we
want.

This continues to be ridiculously fragile, but it is at least working,
for now...