feat(nix): run the Nix installer
I _think_ this can be run non-interactively, but it does need `sudo`, so
trying that here. Note that the install script just downloads a binary
installer which itself is way more complicated and which uses sudo
itself. In my clowny testing, it seems to execute without prompts if it
detects it is running non-interactively.

After it completes, open a new shell and test with:

   nix-shell -p nix-info --run "nix-info -m"