chore(homebrew): switch to Neovim "HEAD"
Because 0.10 is not out yet, but I want to try:

- https://github.com/neovim/neovim/pull/28176 ("feat(comment): add
  built-in commenting")

TBH, surprised something like this would make it into core, but it did,
and the implementation looks well-thought out.

Note that I am not updating my appimage install on Debian, because I
don't want to point at a nightly there (the SHA checksum will change).
I'll update that later.