fix(nvim): remove legacy Command-T config
As the comment indicated, most (all) of these settings have no effect in
the Lua version of Command-T, which is the only one I use now.

Recent Neovim update forces the issue, however, because the `'term'`
option has been removed:

    command-t.lua:13: Invalid option name: 'term'

I could remove just the lines that relate to that option, but I may as
well wipe them all out.