feat(dotfiles): add `git review` alias
This one is perhaps not as glorious as the name would suggest, but it
supports a common workflow here at Datadog, which is to do a
fast/shallow clone of the default branch, and then clone coworkers'
branches on demand with `git review ${branch-name}`.

Could be improved (ie. not hard-coding `origin` remote name but rather
inferring it via a heuristic), and could handle the no-arguments case a
bit better; currently it just blows up with:

    fatal: missing branch name; try -b

(Would be nice to show help in that case.)