feat(dotfiles): add `dog` utility
Continuing in the tradition of `portool` and `gg`, this is `dog`. At the
moment, it doesn't do much at all (just `go` to jump to a tmux session
and `help` to show help). I resurrected the completion from the last
tool, but note that for now there are no commands that take file
arguments; that bit is commented out for now because I don't want to
have to look it up again in the future when I have such a subcommand.