feat(nvim): add mappings for ".unit.tsx?" alternate
The configuration is deeply confusing, even after reading the docs:

- https://github.com/tpope/vim-projectionist/blob/729f3630a4c19e52e1fd825802d2a1659daeaf48/doc/projectionist.txt

Adding a line here helps us navigate from "source.ts" to
"source.unit.ts":

    ['*' .. extension] = {
      ['alternate'] = {
        '{dirname}/{basename}.test' .. extension,
        '{dirname}/{basename}.unit' .. extension,
        '{dirname}/__tests__/{basename}.test' .. extension,
        '{dirname}/__tests__/{basename}-test' .. extension,
        '{dirname}/__tests__/{basename}-mocha' .. extension,
      },
      ['type'] = 'source',
    },

The other way is not so straightforward. My first attempt looked like
this:

    ['*.unit' .. extension] = {
      ['alternate'] = '{basename}' .. extension,
      ['type'] = 'test',
    },

based on the config that I already had for "*.test":

    ['*.test' .. extension] = {
      ['alternate'] = '{basename}' .. extension,
      ['type'] = 'test',
    },

But I think that was wrong and broken. "*.test" is short for
"**/*.test", and when you pull out "{basename}" you're basically
grabbing only the "**" part, not the "*" part. So, that means given
"src/subdir/hey.unit.ts", trying to go back to the source would take you
to "hey.ts".

So, the proper config is:

    ['*.unit' .. extension] = {
      ['alternate'] = '{dirname}/{basename}' .. extension,
      ['type'] = 'test',
    },

TBH, I think this would work too:

    ['*.unit' .. extension] = {
      ['alternate'] = '{}' .. extension,
      ['type'] = 'test',
    },

But for consistency with the others, I am using the former pattern
instead. (I will probably simplify in the next commit though, just to
prove that it does work.)