refactor(nvim): update help tags a different way
`-c` used to hang. Many commands on different lines used to yield an
error. But `-es` plus "piped" (ie. commands in a list separated by a
vertical bar) seems to both work (ie. it produces `tags` files), exit
cleanly (ie. it exits with status 0), and not hang.

Another option I was looking at was using `:cquit 0`, which works
interactively, but not at the end of an `nvim -i NONE -u NONE -es`
invocation with multiple commands:

- https://stackoverflow.com/a/66286688