chore: re-encrypt encrypted files
Some of these get picked up automatically and Git ends up re-"cleaning"
them. But I had to force the others to get updated with:

    git add --renormalize .

I presume this happens because the of the
"diff.git-cipher.cachetextconv" config setting, which causes a cache to
be used that ends up causing some of these items to be stale.