chore(nvim): use `cmp.ConfirmBehavior.Replace`
This is very close to what I want, so going to give it a try.

My only gripe with it is that it sometimes replaces more than I would
want, so I'm hopeful upstream will add something like the `MatchSuffix`
behavior proposed here:

- https://github.com/hrsh7th/nvim-cmp/issues/1716

Most of the time, `Replace` does what you want, but it doesn't in some
rare edge cases, because it greedily replaces everything up to the next
keyword boundary. It's hard to come up with a non-contrived example to
illustrate this, but let me try; let's start with something like this:

      const actions = [select_next_item, select_prev_item];

The cursor is before the `n` of `select_next_item`. You want to
autocomplete that to `select_first_item`, then type ", select_" to
produce:

      const actions = [select_first_item, select_next_item, select_prev_item];

But what you actually get as soon as you hit <C-y> is:

      const actions = [select_first_item, select_prev_item];

because your first completion replaces the entire keyword.