feat(nvim): hack in global command mode aliases
This is an inglorious hack, but... it is useful.

In Zsh I have a global alias that expands anywhere it is used:

    alias -g G="|grep"

I can type:

    tcs G error

to stand for:

    tcs | grep error

I want to do something similar in Neovim, so that I can type:

    :Ack term $PROFILING

and have it expand to:

    :Ack term deep/directory/structure/containing/profiling/code another/deep/hierarchy/with/profiling/stuff

So, we do that, horrendously, by exporting an environment variable with
the paths (encrypted so as not to leak information about internal repo
structure), and then we observe the command line for edits.

This works, although it sucks. Ideally, I'd also have:

1. Autocompletion of environment varibales names.
2. Tab `<Tab>` to expand.
3. Both of the above without any hacks.

but going to live with this as-is for now.