chore: don't version ~/Documents/Personal/Corpus
For complicated reasons I don't want to spell out in this commit
message.

The immediate problem is that I can't add this directory as a
subdirectory of another directory that happens to be a Git repo without
this directory being treated as a submodule, even if I use a separate
git-dir for the inner repo.

So, feck it. Just make it a collection of normal files.