fix(nvim): make `is_only_window` check axis specific
Just noticed that the `is_only_window` check doesn't handle a layout
like this:

     *------|------|------*
     |      |      |      |
     |      |      |      |
     |      |      |      |
     |      |      |      |
     |      |      |      |
     *------|------|------*

when resizing vertically. None of these are "only windows", but none of
them should be allowed to resize vertically. The same is true for these
when resizing horizontally:

     *--------------------*
     |                    |
     *--------------------*
     |                    |
     *--------------------*
     |                    |
     *--------------------*

None of those are "only" windows, but none of them should be allowed to
resize horizontally.