feat(nvim): add a `:CommandTShellbot` command
This makes it easy to get back to previously asked ChatGPT questions.

I am thinking about making this the default workflow for opening
shellbot (instead of `:ChatGPT`). ie. hit a binding, type a question (to
go back to previous session, or start a new session) or just hit enter
to open a blank new session.

I can, for example, insert a prompt in the list and add a custom `open`
implementation to create a new session. What I can't do (at least not
right now) is the Corpus-style workflow of offering autocomplete and
falling back to creating a new document. Not sure whether to hack
Command-T to support this flow, or just implement something like Corpus
does. I'll think about it.