perf: add `bin/benchmark` script
To quickly measure shell startup. This is a bit quicker, although
obviously not as detailed as the profiling stuff I have commented out in
my `.zshrc`.

This is based on a script seen at:

- https://github.com/zimfw/zimfw/wiki/Speed

with the following changes:

1. Look for my RPROMPT instead of the left-side prompt.
2. Print milliseconds instead of microseconds.
3. Timeout in 2s instead of 10s.
4. Ample comments.

I know I could have just asked ChatGPT to explain this if I ever forget
how it works, but I rarely use TCL and I will rarely edit this script,
so I prefer to comment it thoroughly.