perf(dotfiles): make functions autoloaded
Note that for this to work, the functions have to be the only thing in
the file; otherwise, the first invocation will load them, and the second
invocation will run them.

This is why I put the `_jump_complete` and `compctl` calls in the
`functions` file instead of at the end of `functions.d/jump`. I'm not
sure whether this is a good call or not, but will give it a try as-is
for now.