fix(nvim): make split resizing work outside of tmux
Inside tmux, the "<M-" mappings work fine and are triggered by both the
"Option" (⌥) key and the "Command" (⌘) key on macOS.

Outside of tmux, only the "Option" key was working.

The "<D-" mappings work outside of tmux, but not inside it, and they
only work with the Command key.

So, the net effect of defining both "<M-" and "<D-" mappings is that
both keys work both inside and outside of tmux. Personally, I would have
been fine with just the Command key working, but it wasn't to be. I
tried to work around this at the kitty layer, but couldn't get it to
work.

One place where this inconsistency is still visible is outside of
Neovim:

- In tmux, both ⌘-(left/right) and ⌥-(left/right) will move
  word-wise at the shell prompt.
- Outside of tmux, only ⌥-(left/right) does word-wise movement.
  ⌘-(left/right) ends up echoing "D" or "C" respectively. `kitty
  +kitten show_key` indicates it's seeing the following for those
  ⌘-(left/right) key presses:

      UNIX            send_text
    ^[[1;9C         \x1b[1;9C
    ^[[1;9D         \x1b[1;9D

Closes: https://github.com/wincent/wincent/issues/147