feat(dotfiles): improve copy mode navigation in tmux
Suppress the top right indicator in most situations, which eliminates
the confusing from-the-bottom indexing. That in turn means I can turn
search wrapping back on.

As noted in the comments, there are some possible future enhancements,
but for now this gives us a less confusing experience at the cost off
showing an empty bottom pane border on all panes instead of just the
panes that have another pane underneath them. I _could_ suppress the
border and turn it back on when entering copy mode, but that would cause
the pane size to change, and I don't think that's great, so this fills
like the right trade-off for now.