refactor(dotfiles): update word-separators setting
As of tmux 3.3, and specifically commit 022d0210c5afa4e5 ("More
accurate vi(1) word navigation in copy mode and on the status line",
2021-05-13), the default is very comprehensive.

Before that the default was just " ". Previously it had been " -_@" and
I had added ":/" to it, but in tmux commit e90d4a6021a45e4f ("Add
formats for word and line under the mouse...", 2019-05-26) it was
stripped back to " ". So, in the `%else` branch here I am just setting
the whole thing to " -_@:/" instead of appending to it.