fix(nvim): make more plug-ins deactivatable
Needed a couple of `pcall` to gracefully degrade when a `plugin.load()`
is commented out.

Added note to make clear that you can't deactivate nvim-cmp without
affecting the "downstream" plugins that depend on it. Given that those
are all third party, I can't add `pcall` checks to those. I could add a
"DSL" though to make the dependency explicit, but it wouldn't actually
do anything; ie:

    wincent.plugin.load('base', function ()
      wincent.plugin.load('extra')
    end)

or:

    wincent.plugin.load('base').dependents(function ()
      wincent.plugin.load('extra')
    end)

And it would just execute the callback. To disable the base plugin,
you'd have to comment out the entire block. Not going to do it though
because I'm not convinced it's actually worth it.