feat(dotfiles): add Hammerspoon menu item for changing display arrangements
I have a sit-stand standing desk adapter now, so I need to be able to
quickly switch between:

- Standing: External display is the main display, at eye level; internal
  display is immediately underneath the external one, centered, right in
  front of the keyboard (not an ideal place for it, but it's literally
  the only place it can go).
- Sitting: External display is the main display; internal display is off
  to the right on a laptop stand, slightly lower.

Now, with the automatically shown/hidden menu bar, and Ice struggling to
manage all the hidden and "always hidden" menu items, the menu item
itself isn't super reliable. So, I added hotkey bindings as well to
enable me to force an arrangement.

But when it works, the menu item hides itself when the laptop is
unplugged, and reveals itself when it is plugged back in.