feat(nvim): add mini.ai plugin
- https://github.com/echasnovski/mini.ai

Enhances built-in text objects in the following ways:

- Operations involving pairs like:

      vi( (or other bracket-like objects)
      vi" (or other quote-like objects)

  have a "balanced" form. Say you're inside a set of multiply-nested
  parens, you can do something like `vi(` to select the innermost ones,
  then `i(` again to select the next pair, repeating as you like.

- `b` is a shortcut for any bracket-like object (eg. `(`, `{`, `[`).

- `q` is a shortcut for any quote-like object (eg. `'`, `"`, and
  backtick).

You can also define your own objects (I'm not doing that yet), and there
are other mappings that I doubt I'll use but which do exist; eg.

- `g]` and `g[` motions; eg:
  - `g]b` would take you to the right edge of the next bracket-like.
  - `g[q` would take you to the left edge of the previous quote-like.

- ` ` (space) is supposed to be a text object that targets, digits,
  punctuation, or whitespace; in my experience, it didn't work.

- There are `f` (function call) and `a` (argument) text objects, but I
  usually forget about them.

- There is a `?` text object that prompts the user for beginning and end
  characters.

All of this seems to work just fine with vim-surround:

- https://github.com/tpope/vim-surround

ie. I can type something like `ysiq)` to sorround the nearest quote-like
with parentheses.

Note that there is also a "mini.surround" module that I might consider
switching to, just to have more stuff in Lua and less in Vimscript:

- https://github.com/echasnovski/mini.surround