feat: add new color schemes
Replace the (abandoned) theme source with tinted-theming (the unofficial
successor to the canonical Base16 project). This is simpler because
the themes are all centralized in a single repo.

Update the themes using the latest templates (some color adjustments,
some additional colors, many new schemes). Includes commit to
base16-nvim submodule because that got updated too.

* aspects/nvim/files/.config/nvim/pack/bundle/opt/base16-nvim c0c2275...7dcb960 (1):
  > chore: run theme update from super project