refactor: use templates from tinted-themes repos
These are mostly equivalent, but I like the idea of keeping as much
stuff under one roof as I can.

Differences:

- "tinted-kitty" currently doesn't set `cursor_text_color` like the old
  one did; will probably want to cherry-pick this commit:

  https://github.com/kdrag0n/base16-kitty/commit/0898f2677f3a583cc6a89bde29b2b05ac2041e0b

- New repos come with some different config that required me to move
  stuff around:

  - tinted-kitty wants to write templates into separate "colors/" and
    "colors-256/" directories (previously I dumped ".conf" and
    "-256.conf" files into a single "colors/" directory).

  - tinted-shell wants to write into a "scripts/" directory. Given that
    that is a very generic name, I'm wiring things up to write to
    "~/.zsh/colors/scripts/" instead of just "~/.zsh/scripts/", which I
    think would be confusing.

Note that I almost could remove a lot of the builder code at this point
because the new repos include pre-built copies of the output files. Not
doing that yet though because I have a circular dependency of sorts
between the builder script and the base16-nvim repo (which contains a
copy of the output files, but relies on the builder script to build
them).