feat(dotfiles): add `git mine` and `git recent` aliases
Because in the main repo I'm working nowadays I have a lot of branches,
many of them with annoyingly similar or annoyingly long names, making it
hard to find the one I am looking for.

Usually, the one I am looking for is within the most recently modified
ones. So, add two aliases that are useful for listing such branches.

I haven't been using `git sync` lately, which is what I used to use to
address this same problem. I was never entirely happy with that
because it got the clutter out of the way (ie. in the `attic/`
namespace), but it was still clutter. Going to remove it in the next
commit (or perhaps turn it into an alias for `git fresh` or similar seeing as I have muscle memory for it).