refactor(dotfiles): avoid a `cut`
As mentioned in the parent commit, I built the `dupes` pipeline up
iteratively. At the end, it was working, so I committed it, but there
was an avoidable `cut` in there, which I am removing in this commit.

Instead of stripping out the "blob" type field, we just instruct `uniq`
to skip over it and adjust our indices in the final `awk` call to print
the fields that we want (now shifted over by one due to the extra
field).