feat(karabiner): make media keys work like function keys when modifiers pressed
Address long-standing gripe of mine that in order to use my Hammerspoon
bindings like `^⌥⌘-F1` I have to hold `fn` as well, and I only have
that on my laptop keyboard.

With this change, F1 alone will continue to be "brightness down", but F1
plus any modifier(s) will be exactly that: a function key plus the
modifier(s). I don't know if any of those media keys are supposed to do
anything different with modifiers, but if so, they won't do it any more;
but this is no loss to me because I never used that functionality!

I must confess that how this all works is confusing:

- In the system settings under "Function keys" I have "Use F1, F2, etc. keys as standard function keys" turned off.
- I've got Karabiner set, via `fn_function_keys`, to assign media
  actions to all the function keys (most of them defaults, but with some
  overrides highlighted).
- To get Karabiner to do what I want, I have to map from the function
  key codes (ie. `key_code: 'f1'`) and not the consumer key codes (eg.
  `consumer_key_code: 'display_brightness_decrement'` etc).
- I'm not 100% I have to do the left-vs-right modifiers separately, but
  it's one of the approaches I tried as I tested out various things
  (most of them failing utterly) before landing on this working
  approach. I guess it makes sense that it should be that way, so I'm
  going to leave it.

Also note the use of `mandatory` and `optional`: `mandatory` ensures we
only kick in if a modifier is pressed; `optional` lets anything else
through; in the `to` binding we have to put the `mandatory` modifier
back because otherwise it gets stripped.