feat(nvim): don't let virtual text overwrite buffer text
This is the downside of `left_align`; it can overwrite the buffer text
if the window is narrow + the buffer text is long enough + the virtual
text is long enough.

The hack here isn't perfect. We don't know how many diagnostics we're
going to have, so we don't know how much space to reserve for the prefix
characters (one "■" per diagnostic on that line). There is a `prefix`
callback we can implement, and it receives `total` as an argument, but
it gets called _after_ the `format` callback, so we can't stash the
`total` in the `user_data` field and read it out there. We could do
`vim.diagnostic.get()`, but that sounds expensive and slow if we have to
do it for every call to `format()`.

So, we just set a reasonable `padding` and allow that to cover (most of)
our bases. Based on the calculation, we decide whether to truncate the
diagnostic or not.