fix(nvim): make `wrap_mapping()` work again
One of the many things that broke when I updated indent-blankline.nvim
and I didn't notice. `:IndentBlanklineRefresh` got removed:

- https://github.com/lukas-reineke/indent-blankline.nvim/commit/13e415c6f70fed2f37162c2fbfd07b33d144e15b
- https://github.com/lukas-reineke/indent-blankline.nvim/wiki/Migrate-to-version-3

Note that I could have used `debounced_refresh(0)`, but `refresh(0)`
feels better, and I think it's ok to do this synchronously and
immediately in response to a user input that is relatively rare, like
changing a fold.