refactor(nvim): switch to better fix for broken custom source
Removes the hack, and avoids the use of `set_onetime`. As the latter
isn't documented, maybe I shouldn't have been using it anyway. The only
bad thing about this is I might be invalidating the config cache on
every key stroke, so I'll see if I can optimize that in the next commit.