chore: updated tinted themes
Note there have been breaking changes upstream:

- https://github.com/tinted-theming/schemes/pull/34

that required me to change the builder script. Specifically, they added
a "#" prefix before all the values, but the templates still conserved
their own "#" prefixes, so you would wind up with changes like:

    -selection_background #a5a2a2
    -selection_foreground #090300
    +selection_background ##5c5855
    +selection_foreground ##a5a2a2

instead of:

    -selection_background #a5a2a2
    -selection_foreground #090300
    +selection_background #5c5855
    +selection_foreground #a5a2a2

Given that I don't control the templates, hack around this in the
builder script.

The other breaking change is a loss of functionality in the tmux
template:

- https://github.com/tinted-theming/tinted-tmux/pull/24

This means we no longer have a "zoomed" indicator in the pane listing.

Before:

    1:vim*Z   2:zsh-

After:

    1:vim   2:zsh-

* aspects/nvim/files/.config/nvim/pack/bundle/opt/base16-nvim 7dcb960...3126f8c (1):
  > chore: run theme update from super project