feat(nvim): implement equivalent of `cmp.ConfirmBehavior.MatchSuffix`
Not sure if this _exactly_ matches the behavior discussed in:

- https://github.com/hrsh7th/nvim-cmp/issues/1716

but it at least approximates it. When confirming a completion, it
decides at runtime whether to use `Replace` or `Insert` behavior.
Because I already have code to turn off ghost text inside keywords,
ghost text (still) behaves appropriately. Likewise, no special handling
of cursor position is implemented; it lets nvim-cmp do it's thing in
both `Replace` and `Insert` cases.

Reasons for taking this approach? Well, I was looking at the place where
a real `MatchSuffix` feature would be implemented in nvim-cmp:

- https://github.com/hrsh7th/nvim-cmp/blob/b555203ce4bd7ff6192e759af3362f9d217e8c89/lua/cmp/core.lua#L455-L459

and learned a few things about how the existing behaviors are
implemented.

For one thing, it's the LSP server itself that provides the "replace"
and "insert" ranges for the `textEdit` operations, although nvim-cmp
does a lot of massaging of the responses. I was surprised to find this
old ticket describing the then-limitations of the protocol and how
VSCode was effectively sprinkling in its own logic to provide a good
experience:

- https://github.com/Microsoft/language-server-protocol/issues/264

Given that the LSP server itself is the one deciding which characters to
be replaced (or merely inserted), it feels like it's probably desirable
for nvim-cmp to be as unopinionated as possible about how to perform the
edits. That's what led me to investigate the approach taken in this
commit: instead of introducing a new `ConfirmBehavior`, we dynamically
choose the one that suits us best at runtime. The only thing I don't
like about this approach is it might expose me to breaking changes if
the nvim-cmp data types ever change. Specifically, we're relying on
having access to an `entry` object (although we can degrade gracefully
in the absence of one) with these fields:

- `entry.completion_item.textEdit`
- `entry.replace_range.end`
- `entry.context.cursor.col`
- `entry.context.cursor_after_line`
- `completion_item.textEdit.newText`