feat(nvim): make <S-Tab> the inverse of <Tab> in edge case
Old behavior when pressing `<S-Tab>`:

- If nvim-cmp menu is visible, select previous item
- If inside a LuaSnip snippet and eligible to jump to previous
  placeholder, jump to previous placeholder
- Fallback to whatever is normally mapped to `<S-Tab>` (effectively,
  equivalent to inserting another tab).

New behavior - the same as before, but we insert two more cases before
the fallback:

- If in the leading indent, dedent the line. In buffer with
  `'expandtab'` set (which is the case for most of my buffers), the
  difference between `<C-D>` (dedent) and `<BS>` is minor (both will
  delete an entire `'shiftwidth'`), but we're using the former anyway
  because it seems more semantically accurate. In a buffer with
  `'noexpandtab'`, there is no change.
- If at any other whitespace character, call `smart_bs()`. In
  `'expandtab'` buffers this is equivalent to `<BS>`. In `'noexpandtab'`
  buffers it is also going to pretty much correspond to `<BS>`.

The net effect of all that is it makes `<S-Tab>` feel like the inverse
of `<Tab>` in more cases. That is, in the leading indent, hitting
`<Tab>` five times and `<S-Tab>` five times will indent and dedent the
line back to where you started. Likewise, in the middle of a line if you
are inserting whitespace with `<Tab>` then `<S-Tab>` will undo it for
you.

Note that if you use `<S-Tab>` in mid-line whitespace and you get all
the way back to a non-whitespace character, it will _not_ delete that
character because that would feel weird and destructive. Instead, you'll
effectively wind up seeing an nvim-cmp menu (if there is anything to
complete), or hitting the fallback, which will effectively _add_ a
tab. That _is_ a bit weird, but it is going to be relatively rare, and
it's non-destructive, so I think it's the right choice.