fix(nvim): avoid unwanted conceal in Markdown
Not sure if I am going to keep these overrides, but I think it's useful
to have them in the repo history anyway.

I personally think Vim should be for editing Markdown, not viewing it
(if I want to view it without the markup, I use something else; and
besides, the whole point of Markdown is for the source to be readable
anyway).

Since switching to treesitter, the concealing has been driving me crazy;
so, override it. Wondering why `'conceallevel'` is 2 anyway? `:verbose
set conceallevel?` reveals:

  conceallevel=2
        Last set from Lua (run Nvim with -V1 for more details)

which isn't all that helpful. With `-V1`, we discover this:

  conceallevel=2
        Last set from ~/code/wincent/aspects/nvim/files/.config/nvim/lua/wincent/autocmds.lua line 129

So, it's actually me... and specifically this line, which uses 2 as a
default...

  local conceallevel = autocmds.conceallevel_filetypes[filetype] or 2

I can't remember why I did this, but I am guessing it has to do with
the indentLine plug-in that I used to use before indent-blankline.nvim.
The former used conceals, while the latter uses virtual text.

`git blame` doesn't shed any light, so I think in the next commit I'm
going to revert this little exploration and change that `2` to a `0` and
see how things work...